reflection/reflection.go
package main

import (
	"reflect"
)

func walk(x interface{}, fn func(input string)) {
	val := getValue(x)

	walkValue := func(value reflect.Value) {
		walk(value.Interface(), fn)
	}

	switch val.Kind() {
	case reflect.String:
		fn(val.String())
	case reflect.Struct:
		for i := 0; i < val.NumField(); i++ {
			walkValue(val.Field(i))
		}
	case reflect.Slice, reflect.Array:
		for i := 0; i < val.Len(); i++ {
			walkValue(val.Index(i))
		}
	case reflect.Map:
		for _, key := range val.MapKeys() {
			walkValue(val.MapIndex(key))
		}

	}
}

func getValue(i interface{}) (val reflect.Value) {
	val = reflect.ValueOf(i)
	if val.Kind() == reflect.Ptr {
		val = val.Elem()
	}

	return
}

func main() {
	profile := []struct {
		Age  int
		City string
	}{
		{33, "Manila"},
		{20, "Cebu"},
	}

	var got []string
	walk(profile, func(input string) {
		got = append(got, input)
	})
}
